Changes to make components more consumable
