+++ docker-compose.yml
@@ -21,7 +21,7 @@ services:
     container_name: dbzui-db-pg
     image: debezium/example-postgres:1.3
     ports:
-      - "65432:5432"
+      - "5432:5432"
     environment:
       - POSTGRES_USER=postgres
       - POSTGRES_PASSWORD=postgres
