+++ ui/packages/ui/src/app/pages/connectors/ConnectorsTableComponent.css
