+++ ui/packages/ui/src/app/pages/createConnector/index.ts
@@ -1,1 +1,2 @@
+export * from './CreateConnectorComponent';
 export * from './CreateConnectorPage';
